Extract NotificationToolbar
