+++ ArtemisKit/Sources/Notifications/NotificationView.swift
@@ -28,7 +28,7 @@ struct NotificationView: View {
                 } else {
                     List {
                         ForEach(notifications) { notification in
-                            NotificationCell(notification: notification)
+                            NotificationListContent(notification: notification)
                                 .onTapGesture {
                                     dismiss()
                                     guard let type = notification.pushNotificationType,
@@ -60,7 +60,7 @@ struct NotificationView: View {
     }
 }
 
-struct NotificationCell: View {
+private struct NotificationListContent: View {
 
     let notification: Notification
 
@@ -82,62 +82,6 @@ struct NotificationCell: View {
             }
             .padding(.l)
             .cardModifier(backgroundColor: Color.Artemis.modalCardBackgroundColor)
-        } else {
-            EmptyView()
-        }
-    }
-}
-
-struct NotificationBell: ViewModifier {
-
-    @StateObject private var viewModel = NotificationViewModel()
-
-    @State private var isNotificationSheetPresented = false
-
-    func body(content: Content) -> some View {
-        content
-            .toolbar {
-                ToolbarItem(placement: .navigationBarLeading) {
-                    Button {
-                        isNotificationSheetPresented = true
-                    } label: {
-                        Image(systemName: "bell.fill")
-                            .overlay(Badge(count: viewModel.newNotificationCount))
-                    }
-                }
-            }
-            .sheet(isPresented: $isNotificationSheetPresented) {
-                NotificationView(viewModel: viewModel)
-            }
-            .task {
-                await viewModel.subscribeToNotificationUpdates()
-            }
-    }
-}
-
-public extension View {
-    func notificationToolBar() -> some View {
-        modifier(NotificationBell())
-    }
-}
-
-struct Badge: View {
-    let count: Int
-
-    var body: some View {
-        // swiftlint:disable:next empty_count
-        if count > 0 {
-            ZStack(alignment: .topTrailing) {
-                Color.clear
-                Text(String(count))
-                    .font(.system(size: 16))
-                    .foregroundColor(.white)
-                    .padding(.s)
-                    .background(Color.red)
-                    .clipShape(Circle())
-                    .alignmentGuide(.top) { $0[.bottom] }
-                    .alignmentGuide(.trailing) { $0[.trailing] - $0.width * 0.25 }
-            }
         } else {
             EmptyView()
         }
